old_archives/requirements.txt
# Core Dependencies
faster-whisper>=1.2.0
python-dotenv>=0.19.0
schedule>=1.2.2

# Video Processing
yt-dlp>=2023.3.4
opencv-python>=4.5.0
numpy>=1.21.0

# YouTube API
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# HTTP Requests
requests>=2.25.0

# Optional - Not directly used in current code but might be needed for future features
yake>=0.6.0

# System Requirements
# ffmpeg (install via system package manager)
